Refatora provider do participante direto
