+++ src/providers/DirectParticipantProvider/dto/IOrderResponse.ts
@@ -1,0 +1,7 @@
+export default interface OrderResponse {
+  deep_link: string;
+  order_id: string;
+  qr_code: string;
+  qr_code_text: string;
+  status: string;
+}
