+++ src/providers/DirectParticipantProvider/models/IDirectParticipantProvider.ts
@@ -1,4 +1,7 @@
+import Order from '../dto/IOrder';
+import OrderResponse from '../dto/IOrderResponse';
+
 export default interface IDirectParticipantProvider {
-  generateTransaction(order: Order): Promise<string>;
+  generateTransaction(order: Order): Promise<OrderResponse>;
   checkStatus(payload: string): Promise<string>;
 }
